addressing PR asks
