+++ com.microsoft.mrtk.spatialmanipulation/PlacementHub.cs
@@ -1,7 +1,7 @@
-using Microsoft.MixedReality.Toolkit;
-using Microsoft.MixedReality.Toolkit.SpatialManipulation;
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
@@ -11,9 +11,14 @@ namespace Microsoft.MixedReality.Toolkit.SpatialManipulation
     {
         [SerializeReference]
         [InterfaceSelector]
-        public List<ITransformation> transformations =new List<ITransformation>();
+        private List<ITransformation> transformations = new List<ITransformation>();
 
-        (Vector3, Quaternion, Vector3) targetPose;
+        internal List<ITransformation> Transformations => transformations;
+
+        /// <summary>
+        /// The target pose 
+        /// </summary>
+        private (Vector3, Quaternion, Vector3) targetPose;
 
         // Make this it's own class so we can potentially allow for alternate implementations which "blend" transformations or
         // use a transformation's execution order differently.
@@ -365,43 +370,5 @@ namespace Microsoft.MixedReality.Toolkit.SpatialManipulation
             }
         }
         #endregion
-    }
-}
-
-/// <summary>
-/// Interface which describes a transformation that is applied to MixedRealityTransform
-/// </summary>
-public interface ITransformation
-{
-    public (Vector3, Quaternion, Vector3) ApplyTransformation(Vector3 initialPosition, Quaternion initialRotation, Vector3 initialLocalScale);
-
-    /// <summary>
-    /// Execution order priority of this constraint. Lower numbers will be executed before higher numbers.
-    /// </summary>
-    public int ExecutionPriority { get; }
-}
-
-/// <summary>
-/// A transformation which restricts the scale of of the MixedRealityTransform
-/// </summary>
-public class MinMaxConstraintTransformation: ITransformation
-{
-    public Vector3 minScale = Vector3.one * 0.2f;
-
-    public Vector3 maxScale = Vector3.one * 2.0f;
-
-    protected const int scale_priority = -1000;
-    protected const int rotation_priority = -1000;
-    protected const int position_priority = -1000;
-    protected const int constraint_priority_modifier = 1;
-
-    public int ExecutionPriority => throw new NotImplementedException();
-
-    public (Vector3, Quaternion, Vector3) ApplyTransformation(Vector3 initialPosition, Quaternion initialRotation, Vector3 initialLocalScale)
-    {
-        initialLocalScale.x = Mathf.Clamp(initialLocalScale.x, minScale.x, maxScale.x);
-        initialLocalScale.y = Mathf.Clamp(initialLocalScale.y, minScale.y, maxScale.y);
-        initialLocalScale.z = Mathf.Clamp(initialLocalScale.z, minScale.z, maxScale.z);
-        return (initialPosition, initialRotation, initialLocalScale);
     }
 }
