+++ com.microsoft.mrtk.spatialmanipulation/Transformations.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a87b6c4bbf39bc744b68022807cb8e33
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
