+++ com.microsoft.mrtk.spatialmanipulation/Transformations/ITransformation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 35919b876a9374741bf9babaede7ffdd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
