+++ com.microsoft.mrtk.spatialmanipulation/Transformations/MinMaxConstraintTransformation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bdde46f9660b4d744936a19cb3fc28a9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
